fix: grammar errors in History and Kavaleridze, remove mobile sub menu on Header
