+++ src/assets/tempData/staticInfo/movie.ts
@@ -8,7 +8,7 @@ export const movieData: IPage = {
     {
       type: 'text',
       textContent:
-        'Як кінорежисер, Кавалерідзе створив лише 9 повнометражних фільмів. В кіно Кавалерідзе прийшов після знайомства з Олександром Довженком і заступником голови ВУФКУ Зіновієм Сидерським. Його кінематографічним дебютом став фільм “Злива” 1929 року, який було втрачено при загадкових обставинах. Наступний фільм “Перекоп” 1930 р., близький до епічного кіно, третій — побудований на монтажних прийомах “Штурмові ночі”, 1931 р. та один з перших звукових фільмів “Коліївщина” 1933 р. Ці роботи не знаходять  підтримки у радянських кінокритиків.',
+        'Як кінорежисер, Кавалерідзе створив лише 9 повнометражних фільмів. В кіно Кавалерідзе прийшов після знайомства з Олександром Довженком і заступником голови ВУФКУ Зіновієм Сідерським. Його кінематографічним дебютом став фільм “Злива” 1929 року, який було втрачено при загадкових обставинах. Наступний фільм “Перекоп” 1930 р., близький до епічного кіно, третій — побудований на монтажних прийомах “Штурмові ночі”, 1931 р. та один з перших звукових фільмів “Коліївщина” 1933 р. Ці роботи не знаходять  підтримки у радянських кінокритиків.',
       columns: 1,
     },
     {
