+++ src/assets/tempData/staticInfo/museumHistory.ts
@@ -42,7 +42,7 @@ export const museumData: IPage = {
         {
           imageLink: slider6,
           imageLabel:
-            'У музеї демонструються проекти пам‘ятників, станкові скульптури, які зображують Тараса Шевченка, Григорія Сковороду, Лесю Українку, Миколи Амосова та інших видатних людей.',
+            'У музеї демонструються проекти пам‘ятників, станкові скульптури, які зображують Тараса Шевченка, Григорія Сковороду, Лесю Українку, Миколу Амосова та інших видатних людей.',
         },
         { imageLink: slider7, imageLabel: 'Для огляду постійно відкрита меморіальна кімната — кабінет Івана Петровича.' },
         { imageLink: slider8, imageLabel: 'До уваги відвідувачів — оглядова екскурсія, відеозаписи кінофільмів І. Кавалерідзе.' },
@@ -56,7 +56,7 @@ export const museumData: IPage = {
     {
       type: 'text',
       textContent:
-        'Іван Кавалерідзе — видатна особистість, талановитий скульптор, великий кінорежисер і драматург. Він безумовно геній минулого століття, “український Мікеланджело”, як його називали. ',
+        'Іван Кавалерідзе — видатна особистість, талановитий скульптор, великий кінорежисер і драматург. Він, безумовно, геній минулого століття, “український Мікеланджело”, як його називали. ',
       columns: 1,
     },
     {
@@ -84,7 +84,7 @@ export const museumData: IPage = {
     {
       type: 'text',
       textContent:
-        'Кавалерідзе створював “пластичні” фільми, образи яких можна описати як кутасті, кубічно чіткі, або рвані та бурхливі. Його скульптури — глибокі та майстерні, з відчуттям авторської душі та “думки серцем”.',
+        'Кавалерідзе створював “пластичні” фільми, образи яких можна описати як кутасті, кубічно-чіткі, або рвані та бурхливі. Його скульптури — глибокі та майстерні, з відчуттям авторської душі та “думки серцем”.',
       columns: 1,
     },
     {
