+++ src/components/Header/parts/NavMenu.tsx
@@ -1,8 +1,22 @@
-import { FC, PropsWithChildren } from 'react';
+import { FC } from 'react';
+import { List, useTheme, useMediaQuery } from '@mui/material';
+import NavMenuItem from './NavMenuItem';
+import data from '../../../assets/siteData';
 
-import { List } from '@mui/material';
+const {
+  menuList: { main },
+} = data;
 
-const NavMenu: FC<PropsWithChildren> = ({ children }) => {
+interface NavMenuProps {
+  onCloseMobileMenu?: () => void;
+}
+
+const NavMenu: FC<NavMenuProps> = ({ onCloseMobileMenu }) => {
+  const theme = useTheme();
+  const isDesktop = useMediaQuery(theme.breakpoints.up('lg'));
+
+  const navMenu = isDesktop ? main.filter((item) => item.title !== 'Головна') : main;
+
   return (
     <List
       disablePadding
@@ -11,7 +25,9 @@ const NavMenu: FC<PropsWithChildren> = ({ children }) => {
         gap: { xs: 2, md: 3, lg: '60px' },
         gridTemplate: { xs: 'repeat(5, auto)/1fr', lg: '1fr/repeat(4, auto)' },
       }}>
-      {children}
+      {navMenu.map(({ href, title }) => (
+        <NavMenuItem href={href} title={title} key={title} onCloseMobileMenu={onCloseMobileMenu} />
+      ))}
     </List>
   );
 };
