+++ src/components/Kavaleridze/Kavaleridze.tsx
@@ -15,6 +15,7 @@ const Kavaleridze: FC = () => {
     if (location.pathname === '/about') {
       navigate('/about/artist');
     }
+
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [location]);
